Move Searcher out of BiomeSearcher
Remove x,z search params combine into searchRadius.
