+++ src/main/java/sassa/gui/fxmlController.java
@@ -17,6 +17,7 @@ import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
 import org.json.simple.parser.ParseException;
 import sassa.main.BiomeSearcher;
+import sassa.main.Searcher;
 import sassa.util.Singleton;
 import sassa.util.Util;
 import sassa.util.Version;
@@ -40,8 +41,9 @@ public class fxmlController implements Initializable {
     private static long elapsedTime;
 
     static Thread t;
+    static Thread t2;
     static boolean allowThreadToSearch = true;
-    static BiomeSearcher r;
+    static Searcher r;
 
     public static String minecraftVersion = Version.V1_15_2;
     String[] versions = {
@@ -88,10 +90,7 @@ public class fxmlController implements Initializable {
     private TextField seedsToFind;
 
     @FXML
-    private TextField searchX;
-
-    @FXML
-    private TextField searchZ;
+    private TextField searchRadius;
 
     @FXML
     private CheckBox devMode;
@@ -289,11 +288,10 @@ public class fxmlController implements Initializable {
 
     };
 
-    BiomeSearcher createNewThread() throws IOException, FormatException, MinecraftInterfaceCreationException {
-        r = new BiomeSearcher(
+    Searcher createNewThread() throws IOException, FormatException, MinecraftInterfaceCreationException {
+        r = new Searcher(
                 minecraftVersion,
-                Integer.parseInt(searchX.getText()),
-                Integer.parseInt(searchZ.getText()),
+                Integer.parseInt(searchRadius.getText()),
                 Integer.parseInt(seedsToFind.getText()),
                 Long.parseLong(minSeed.getText()),
                 Long.parseLong(maxSeed.getText()),
@@ -359,8 +357,7 @@ public class fxmlController implements Initializable {
 
     private void start() throws IOException, FormatException, MinecraftInterfaceCreationException {
         startBtn.setText("Stop");
-        searchX.setEditable(false);
-        searchZ.setEditable(false);
+        searchRadius.setEditable(false);
         seedsToFind.setEditable(false);
         startTime = System.currentTimeMillis();
         elapsedTime = System.currentTimeMillis();
@@ -368,11 +365,12 @@ public class fxmlController implements Initializable {
         initTimer();
         t = new Thread(createNewThread());
         t.start();
+//        t2 = new Thread(createNewThread());
+//        t2.start();
     }
 
     public void stop() throws InterruptedException, IOException, FormatException, MinecraftInterfaceCreationException {
-        searchX.setEditable(true);
-        searchZ.setEditable(true);
+        searchRadius.setEditable(true);
         seedsToFind.setEditable(true);
         startBtn.setText("Start");
         pauseBtn.setText("Pause");
